Login and Sign up pages front-end and backend modules - first phase.

+++ app-server-backend/users/users-dao.js
@@ -1,36 +1,15 @@
-import userModel from "./users-model";
+import userModel from "./users-model.js";
 
-export const createUser = (user) => {
-    userModel.create(user);
-}
+export const createUser = (user) => userModel.create(user);
 
-export const register = async (user) => {
-    const exisitingUser = await findByUsername(user.username);
-    if (exisitingUser) {
+export const findAllUsers = () => userModel.find();
 
-    }
-} 
+export const findUsersById = (uid) => userModel.findById(uid);
 
-export const findAllUsers = () => {
-    userModel.find()
-}
+export const findByUsername = (username) => userModel.findOne({username})
 
-export const findUsersById = (uid) => {
-    userModel.findById(uid)
-}
+export const findByCredentials = (username, password) => userModel.findOne({username, password}, {password: false});
 
-export const findByUsername = (username) => {
-    userModel.findOne({username})
-}
+export const deleteUser = (uid) => userModel.deleteOne({_id : uid});
 
-export const findByCredentials = (username, password) => {
-    userModel.findOne({username, password}, {password: false});
-}
-
-export const deleteUser = (uid) => {
-    userModel.deleteOne({_id : uid})
-}
-
-export const updateUser = (uid, userUpdates) => {
-    userModel.updateOne({_id: uid}, {$set: userUpdates})
-}
+export const updateUser = (uid, userUpdates) => userModel.updateOne({_id: uid}, {$set: userUpdates})
